# contracts/walksafe.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "WalkSafeAccountGeneration": {
        "description": "Positive account generation bound by the gateway actor assertion.",
        "in": "header",
        "name": "X-WalkSafe-Account-Generation",
        "type": "apiKey"
      },
      "WalkSafeActorAssertion": {
        "description": "Short-lived gateway assertion binding the actor to its authenticated role.",
        "in": "header",
        "name": "X-WalkSafe-Actor-Assertion",
        "type": "apiKey"
      },
      "WalkSafeActorId": {
        "description": "Named actor for secured, rate-limited, or audited operations.",
        "in": "header",
        "name": "X-WalkSafe-Actor-Id",
        "type": "apiKey"
      },
      "WalkSafeAdminAppKind": {
        "description": "Required exact value: ADMIN_ANDROID.",
        "in": "header",
        "name": "X-WalkSafe-App-Kind",
        "type": "apiKey",
        "x-walksafe-required-value": "ADMIN_ANDROID"
      },
      "WalkSafeAdminAudience": {
        "description": "Required exact value: walksafe-admin-api.",
        "in": "header",
        "name": "X-WalkSafe-Audience",
        "type": "apiKey",
        "x-walksafe-required-value": "walksafe-admin-api"
      },
      "WalkSafeAdminBearer": {
        "bearerFormat": "opaque per-device session",
        "description": "Database-backed administrator session; static administrator tokens are rejected.",
        "scheme": "bearer",
        "type": "http"
      },
      "WalkSafeAdminDeviceChallengeId": {
        "description": "Canonical UUID of the fresh one-use device challenge.",
...
  },
  "info": {
    "title": "WalkSafe Assist API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/uploads/{filename}": {
      "get": {
        "operationId": "get_upload_uploads__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-WalkSafe-Original-Access-Grant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Walksafe-Original-Access-Grant"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WalkSafeAdminAppKind": [],
            "WalkSafeAdminAudience": [],
            "WalkSafeAdminBearer": [],
            "WalkSafeAdminDeviceId": [],
            "WalkSafeAdminRole": [],
            "WalkSafeOriginalAccessGrant": []
          }
        ],
        "summary": "Get Upload",
        "x-walksafe-actor-identity-required": false,
        "x-walksafe-actor-identity-source": "admin-session",
        "x-walksafe-required-role": "admin"
      }
    }
  },
...
}